+++ experimental/dds/tree2/src/index.ts
@@ -259,6 +259,14 @@ export {
 	TreeFieldFromImplicitField,
 	TreeNodeEvents,
 	NodeFromSchema,
+	TreeMapNode,
+	InsertableTreeNodeFromImplicitAllowedTypes,
+
+	// experimental @alpha APIs:
+	adaptEnum,
+	enumFromStrings,
+	singletonSchema,
+	typedObjectValues,
 
 	// test recursive schema for checking that d.ts files handles schema correctly
 	test_RecursiveObject,