+++ experimental/dds/tree2/src/test/feature-libraries/sequence-field/sequenceFieldCodecs.spec.ts
@@ -7,9 +7,11 @@ import { mintRevisionTag } from "../../../core";
 import { SequenceField as SF } from "../../../feature-libraries";
 // eslint-disable-next-line import/no-internal-modules
 import { Changeset } from "../../../feature-libraries/sequence-field";
+import { RevisionTagCodec } from "../../../shared-tree-core";
 import { brand } from "../../../util";
 import { TestChange } from "../../testChange";
 import { EncodingTestData, makeEncodingTestSuite } from "../../utils";
+import { populatedMarks } from "./populatedMarks";
 import { ChangeMaker as Change, cases } from "./testEdits";
 
 const encodingTestData: EncodingTestData<Changeset<TestChange>, unknown> = {
@@ -20,13 +22,17 @@ const encodingTestData: EncodingTestData<Changeset<TestChange>, unknown> = {
 			"with repair data",
 			Change.revive(0, 1, { revision: mintRevisionTag(), localId: brand(10) }),
 		],
-		// TODO: Include revive case here or in other encode/decode tests in this file.
-		// It's likely we need a different notion of equality, as revive involves a ProtoNode type
-		// and deep equality of that test case fails on comparing two `StackCursor`s.
-		...Object.entries(cases).filter(([key]) => key !== "revive"),
+		...Object.entries(cases),
+		...populatedMarks.map((mark): [string, Changeset<TestChange>] => [
+			"type" in mark ? mark.type : "NoOp",
+			[mark],
+		]),
 	],
 };
 
 describe("SequenceField encoding", () => {
-	makeEncodingTestSuite(SF.sequenceFieldChangeCodecFactory(TestChange.codec), encodingTestData);
+	makeEncodingTestSuite(
+		SF.sequenceFieldChangeCodecFactory(TestChange.codec, new RevisionTagCodec()),
+		encodingTestData,
+	);
 });