+++ packages/dds/merge-tree/src/opBuilder.ts
@@ -2,17 +2,18 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-/* eslint-disable import/no-deprecated */
 
 import { ISegment, Marker } from "./mergeTreeNodes";
 import {
-	ICombiningOp,
 	IMergeTreeAnnotateMsg,
+	// eslint-disable-next-line import/no-deprecated
 	IMergeTreeGroupMsg,
 	IMergeTreeInsertMsg,
 	IMergeTreeRemoveMsg,
 	MergeTreeDeltaType,
 	IMergeTreeDeltaOp,
+	// eslint-disable-next-line import/no-deprecated
+	IMergeTreeObliterateMsg,
 } from "./ops";
 import { PropertySet } from "./properties";
 
@@ -20,16 +21,13 @@ import { PropertySet } from "./properties";
  * Creates the op for annotating the markers with the provided properties
  * @param marker - The marker to annotate
  * @param props - The properties to annotate the marker with
- * @param combiningOp - Optional. Specifies how to combine values for the property, such as "incr" for increment.
  * @returns The annotate op
  *
- * @deprecated This functionality was not meant to be exported and will be removed in a future release
  * @internal
  */
 export function createAnnotateMarkerOp(
 	marker: Marker,
 	props: PropertySet,
-	combiningOp?: ICombiningOp,
 ): IMergeTreeAnnotateMsg | undefined {
 	const id = marker.getId();
 	if (!id) {
@@ -37,7 +35,6 @@ export function createAnnotateMarkerOp(
 	}
 
 	return {
-		combiningOp,
 		props,
 		relativePos1: { id, before: true },
 		relativePos2: { id },
@@ -50,20 +47,16 @@ export function createAnnotateMarkerOp(
  * @param start - The inclusive start position of the range to annotate
  * @param end - The exclusive end position of the range to annotate
  * @param props - The properties to annotate the range with
- * @param combiningOp - Optional. Specifies how to combine values for the property, such as "incr" for increment.
  * @returns The annotate op
  *
- * @deprecated This functionality was not meant to be exported and will be removed in a future release
  * @internal
  */
 export function createAnnotateRangeOp(
 	start: number,
 	end: number,
 	props: PropertySet,
-	combiningOp: ICombiningOp | undefined,
 ): IMergeTreeAnnotateMsg {
 	return {
-		combiningOp,
 		pos1: start,
 		pos2: end,
 		props,
@@ -77,7 +70,6 @@ export function createAnnotateRangeOp(
  * @param start - The inclusive start of the range to remove
  * @param end - The exclusive end of the range to remove
  *
- * @deprecated This functionality was not meant to be exported and will be removed in a future release
  * @internal
  */
 export function createRemoveRangeOp(start: number, end: number): IMergeTreeRemoveMsg {
@@ -89,11 +81,27 @@ export function createRemoveRangeOp(start: number, end: number): IMergeTreeRemov
 }
 
 /**
+ * Creates the op to obliterate a range
+ *
+ * @param start - The inclusive start of the range to obliterate
+ * @param end - The exclusive end of the range to obliterate
+ *
+ * @internal
+ */
+// eslint-disable-next-line import/no-deprecated
+export function createObliterateRangeOp(start: number, end: number): IMergeTreeObliterateMsg {
+	return {
+		pos1: start,
+		pos2: end,
+		type: MergeTreeDeltaType.OBLITERATE,
+	};
+}
+
+/**
  *
  * @param pos - The position to insert the segment at
  * @param segment - The segment to insert
  *
- * @deprecated This functionality was not meant to be exported and will be removed in a future release
  * @internal
  */
 export function createInsertSegmentOp(pos: number, segment: ISegment): IMergeTreeInsertMsg {
@@ -101,7 +109,6 @@ export function createInsertSegmentOp(pos: number, segment: ISegment): IMergeTre
 }
 
 /**
- * @deprecated This functionality was not meant to be exported and will be removed in a future release
  * @internal
  */
 export function createInsertOp(pos: number, segSpec: any): IMergeTreeInsertMsg {
@@ -116,9 +123,14 @@ export function createInsertOp(pos: number, segSpec: any): IMergeTreeInsertMsg {
  *
  * @param ops - The ops to group
  *
+ * @deprecated The ability to create group ops will be removed in an upcoming
+ * release, as group ops are redundant with he native batching capabilities of
+ * the runtime
+ *
  * @deprecated The ability to create group ops will be removed in an upcoming release, as group ops are redundant with he native batching capabilities of the runtime
  * @internal
  */
+// eslint-disable-next-line import/no-deprecated
 export function createGroupOp(...ops: IMergeTreeDeltaOp[]): IMergeTreeGroupMsg {
 	return {
 		ops,