+++ packages/runtime/runtime-definitions/src/index.ts
@@ -65,16 +65,52 @@ export {
 	SummarizeInternalFn,
 	totalBlobSizePropertyName,
 } from "./summary";
+
+// Re-exports for backwards compatibility.
+// Will be removed in the future.
 export {
-	IIdCompressorCore,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	IdCompressor,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
 	IIdCompressor,
-	SerializedIdCompressor,
-	SerializedIdCompressorWithOngoingSession,
-	SerializedIdCompressorWithNoSession,
-	SessionSpaceCompressedId,
-	OpSpaceCompressedId,
-	SessionId,
-	StableId,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	IIdCompressorCore,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
 	IdCreationRange,
-	initialClusterCapacity,
-} from "./id-compressor";
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	OpSpaceCompressedId,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	SerializedIdCompressor,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	SerializedIdCompressorWithNoSession,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	SerializedIdCompressorWithOngoingSession,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	SessionId,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	SessionSpaceCompressedId,
+	/**
+	 * @deprecated Import from `@fluidframework/id-compressor` instead.
+	 */
+	StableId,
+} from "@fluidframework/id-compressor";